// my-openapi-zod-app/src/api/Schema/ZQueryCategoryModelFilter.ts
//
// This file is generated. Do not edit manually.
//
// @applicationName: GreenOnion.API
// @applicationDescription: GreenOnion.API - 1.0.0.0
// @applicationVersion: 1.0.0.0
// @generatedDate: 2025/07/26
//

import { z } from "zod";
import { ZFilterParameter } from "./ZFilterParameter";

export const ZQueryCategoryModelFilter = z.object({
    categoryId:ZFilterParameter.optional(),
    categoryName:ZFilterParameter.optional(),
    categoryCode:ZFilterParameter.optional(),
    isActive:ZFilterParameter.optional(),
    createdOn:ZFilterParameter.optional(),
    createdById:ZFilterParameter.optional(),
    createdBy:ZFilterParameter.optional(),
    updatedOn:ZFilterParameter.optional(),
    updatedById:ZFilterParameter.optional(),
    updatedBy:ZFilterParameter.optional(),
});